Moved all the previous changes into a new library
lib has two packages, update_package and validate_package. This changes the imports to import lib.update_package.
It keeps the new changes much more separated from the rest. The plan is to gradually move the rest of the project into the new library. This is not possible yet as libraries are supposed to be inside src directories so that they cannot be imported if they are not installed. But moving all of utils into src would have definitely broken the validate pipeline.